HTML Parser
A simple HTML Parser which be used to scan the html files and extract
only words from them and ignore all the tags and other styling elements
present inside the HTML file.

This program will accept only one file at a time and output the words
into a text file. This text file in turn can be used as an input to
stemmers such as Porter or Krovetz. (That was my original intention to
write this program)